Replace use of lsb-release
`lsb-release` is no longer a base package in upcoming Ubuntu versions and was removed from current base images too. This posed a problem when attempting to rebuild this repo's image from scratch.

This commit runs an alternative to `lsb-release` to obtain the ubuntu version name, to fetch the appropriate postgres deb package.